Feat: Add storage and CSV export button.
Track mined tiles in tampermonkey/greasemonkey storage, and add buttons to export to CSV for analysis or reset the logs. Only tracks tiles that were mined while the page is open, and not already mined tiles. Will need changes to use an alternate storage API e.g. chrome storage, in order to work as an extension.